+++ app/scripts/components/corpus-distribution-chart.ts
@@ -1,18 +1,15 @@
 /** @format */
-import angular, { IController, IRootScopeService } from "angular"
+import angular, { type IController, type IRootScopeService, type IScope } from "angular"
 import { Chart } from "chart.js"
 import { html } from "@/util"
-
-const defaultMode: Mode = "relative"
+import { loc } from "@/i18n"
+import { type Option } from "@/components/radio-list"
 
 angular.module("korpApp").component("corpusDistributionChart", {
     template: html`
         <div class="flex flex-col gap-2 items-center">
             <canvas id="distribution-chart"></canvas>
-            <div id="statistics_switch">
-                <a data-mode="relative" href="javascript:"> {{ 'statstable_relfigures' | loc }} </a>
-                <a data-mode="absolute" href="javascript:"> {{ 'statstable_absfigures' | loc }} </a>
-            </div>
+            <radio-list options="$ctrl.modeOptions" ng-model="mode"></radio-list>
         </div>
     `,
     bindings: {
@@ -20,18 +17,30 @@ angular.module("korpApp").component("corpusDistributionChart", {
     },
     controller: [
         "$rootScope",
-        function ($rootScope: IRootScopeService) {
+        "$scope",
+        function ($rootScope: IRootScopeService, $scope: CorpusDistributionChartScope) {
             const $ctrl = this as CorpusDistributionChartController
+            $ctrl.modeOptions = [
+                {
+                    value: "relative",
+                    label: loc("statstable_relfigures", $rootScope["lang"]),
+                },
+                {
+                    value: "absolute",
+                    label: loc("statstable_absfigures", $rootScope["lang"]),
+                },
+            ]
+            $scope.mode = "relative"
             let chart: Chart<"pie">
 
-            const getValues = (mode: Mode) => $ctrl.row.map((corpus) => corpus.values[mode == "relative" ? 1 : 0])
+            const getValues = () => $ctrl.row.map((corpus) => corpus.values[$scope.mode == "relative" ? 1 : 0])
 
             $ctrl.$onInit = () => {
                 chart = new Chart("distribution-chart", {
                     type: "pie",
                     data: {
                         labels: $ctrl.row.map((corpus) => corpus.title),
-                        datasets: [{ data: getValues(defaultMode) }],
+                        datasets: [{ data: getValues() }],
                     },
                     options: {
                         locale: $rootScope["lang"],
@@ -42,24 +51,23 @@ angular.module("korpApp").component("corpusDistributionChart", {
                         },
                     },
                 })
-
-                setTimeout(() => {
-                    const radioList = ($("#statistics_switch") as any).radioList({
-                        selected: defaultMode,
-                        change: () => {
-                            const mode = radioList.radioList("getSelected").attr("data-mode")
-                            chart.data.datasets[0].data = getValues(mode)
-                            chart.update()
-                        },
-                    })
-                })
             }
+
+            $scope.$watch("mode", () => {
+                chart.data.datasets[0].data = getValues()
+                chart.update()
+            })
         },
     ],
 })
 
+type CorpusDistributionChartScope = IScope & {
+    mode: Mode
+}
+
 type CorpusDistributionChartController = IController & {
     row: { title: string; values: [number, number] }[]
+    modeOptions: Option<Mode>[]
 }
 
 type Mode = "relative" | "absolute"