+++ CHANGELOG.md
@@ -5,10 +5,14 @@
 ### Changed
 
 - Replaced Raphael library with Chart.js, used in the pie chart over corpus distribution in statistics
+
+### Refactoring
+
 - Removed the global `c` alias for `console`
 - Removed global `lang`, use `$rootScope["lang"]` instead (outside Angular: `getService("$rootScope")["lang"]`)
 - Removed global `loc_data`, use `$rootScope["loc_data"]` instead (outside Angular: `getService("$rootScope")["loc_data"]`)
 - Removed globals `CSV` and `moment`, import the libraries instead
+- Converted the "radioList" JQuery widget to a component
 
 ### Fixed